Updates to example notebooks and started spatial activations
Updates to example notebooks continues.
Started looking into idea of spatial activations, combining neurons from each channel to understand the types of features the network is activating on for a given image.
NOTE: Activation maps look like they could be the solution, just need to work out how to give the correct layer for it to work.